+++ src/main/java/com/binar/kelompokd/interfaces/IKostReviewService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.UUID;
 
 public interface IKostReviewService {
-  void addReviewKost(UUID kostId, Long userId, Integer rating, String reviewText);
+  void addReviewKost(UUID kostId, Long userId, double rating, String reviewText);
   List<KostReview> getReviewByKostId(UUID kostId);
   void calculateMeanKostRating(UUID kostId);
 }
